backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
passlib[bcrypt]
pydantic
pydantic-settings
python-multipart
python-jose
fpdf2
python-dotenv
pandas
numpy
matplotlib
joblib
scikit-learn
qrcode[pil]
